Print HTTP errors even if they ain't in JSON
Despite the fact API should wrap errors in JSON, we can't guarantee it's
JSON all the time. It some cases we forget something to wrap, in others -
we don't own the errors (e.g. bad gateway between reverse proxy and app
server). Nevertheless, CLI application should be ready for this and
print raw error message in cases it's not JSON.