Use requests.Session() for doing HTTP requests
There are two reasons why using session is a good idea here. The first
one is, session uses a pool with keep-alive connections so any further
requests to the same domain won't initiate a new connection. The second
reason is, it's configurable so we can setup, for example, common headers
here and do not pass them across the code.