Use 'baseurl' app property as API endpoint
So far we setup API host and port through ENV variables and they are
defined globally in main.py. It's not convenient in case when we want
to introduce new commands in separate modules or if we want setup them
through applications settings (not there yet).

In order to avoid those inconveniences, let's define a single source of
truth for any API calls - 'baseurl' property on application instance
that contains and API endpoint in format: "http://{host}:{port}".