Refactor cluster discovering command
There are few good things done in this commit. The first one is that
the command implementation has been moved into separate python module.
The second, new endpoint has been used instead of old ones. This
endpoint follows modern RESTful design as well as allows to pass
arbitrary parameters to discovery driver. Hence, we are not tied to
builtin openstack discovery driver and are free to use anything else.