Make driver argument optional in upgrade case
The driver argument must be always the case for starting upgrade
procedure as operators must decide themselves which driver to use.
However, since Kostyor supports discovery mechanism and it's tied
to deployment the same way as upgrade drivers do, it might be a good
idea to consider driver that's used for discovery to be an upgrade
driver if it's not passed explicitly. Obviously, it's not implemented
so far but let's look in the future and make it happen.

Also, this commit unblocks integration gate of Kostyor master since
it doesn't pass driver when doing an upgrade call to Kostyor CLI.