Propagate parameters to upgrade driver
So far there's no way to propagate any parameters to upgrade drivers
from API (or CLI), that's why such drivers as kostyor-openstack-ansible
requires to hardcode path to OpenStack Ansible root.

However, there's an activity to implement custom parameters support on
backend side, and this is a patch that adds support of client side.