Add support of new response format of `/clusters`
Kostyor's API endpoints `/clusters` is used to respond with the
following JSON:

    {
        "clusters": [
            { ... },
            { ... },
            { ... }
        ]
    }

This format contains a redundant key `clusters`, as it does not carry
any sense. This commit introduces changes that ignores this key and
expects a JSON array as an endpoint response. I.e.:

    [
        { ... },
        { ... },
        { ... }
    ]

The compatibility with old API is preserved.